+++ nautilus_trader/system/config.py
@@ -74,9 +74,9 @@ class NautilusKernelConfig(NautilusConfig, frozen=True):
         The execution algorithm configurations for the kernel.
     controller : ImportableControllerConfig, optional
         The trader controller for the kernel.
-    load_state : bool, default True
+    load_state : bool, default False
         If trading strategy state should be loaded from the database on start.
-    save_state : bool, default True
+    save_state : bool, default False
         If trading strategy state should be saved to the database on stop.
     loop_debug : bool, default False
         If the asyncio event loop should be in debug mode.